src/_3_java_core/collection/lesson_4/workshop/warehouse/InMemoryStorage.java
package _3_java_core.collection.lesson_4.workshop.warehouse;

import _3_java_core.collection.lesson_4.workshop.warehouse.exceptions.ItemNotFoundException;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

public class InMemoryStorage implements Storage {

    public static final int INITIAL_CAPACITY = 256;

    private final Map<String, Wheel> items = new HashMap<>(INITIAL_CAPACITY);
    @Override
    public void putItem(Wheel wheel) {
        items.put(wheel.id(), wheel);
    }

    @Override
    public Wheel getItem(String id) throws ItemNotFoundException {
        Wheel wheel = items.get(id);
        if (wheel == null) {
            throw new ItemNotFoundException(id);
        }
        return items.get(id);
    }

    @Override
    public boolean containsItem(String id) {
        return items.containsKey(id);
    }

    @Override
    public Wheel removeItem(String id) throws ItemNotFoundException {
        Wheel remove = items.remove(id);
        if (remove == null) {
            throw new ItemNotFoundException(id);
        }
        return remove;
    }

    @Override
    public void putAllItems(List<Wheel> items) {
        for (Wheel item : items) {
            putItem(item);
        }
    }
}
